[F-9] move the interface
