+++ src/api/lambda/type.go
@@ -1,0 +1,7 @@
+package apilambda
+
+type WhitelistRequest struct {
+	OldURL     string `json:"oldurl"`
+	NewURL     string `json:"newurl"`
+	MerchantID string `json:"merchantid"`
+}
